+++ src/main.jsx
@@ -2,7 +2,7 @@ import ReactDOM from "react-dom/client";
 import App from "./App.jsx";
 import "./index.css";
 import { Provider } from "react-redux";
-import store from "./store/index.js";
+import store from './store/index.js';
 
 ReactDOM.createRoot(document.getElementById("root")).render(
   <Provider store={store}>
